+++ Lab/it1/src/main/java/org/it1_client_domain/Control_ClientLogin.java
@@ -1,5 +1,10 @@
-package org.client_domain;
-
+package org.it1_client_domain;
+/****
+ * 
+ * @author Elena & Luis Javier
+ * @version 0.1.0
+ *
+ */
 public class Control_ClientLogin {
 
 	public boolean verifyCredentials() {
