+++ Lab/it1/src/main/java/org/it1_presentation/IU_ClientLogin.java
@@ -1,0 +1,10 @@
+package org.it1_presentation;
+
+/****
+ * 
+ * @author Elena & Luis Javier
+ * @version 0.1.0
+ *
+ */
+public class IU_ClientLogin {
+}
