--- Lab/it1/src/main/java/org/presentation/IU_ClientLogin.java
@@ -1,4 +1,0 @@
-package org.presentation;
-
-public class IU_ClientLogin {
-}
