--- Lab/it1_server/src/main/java/org/server_domain/Person.java
@@ -1,9 +1,0 @@
-package org.server_domain;
-
-public class Person {
-
-	private String userName;
-	private String pass;
-	private String type;
-
-}
